update open logic
