+++ Src/WitsmlExplorer.Frontend/components/ContextMenus/ContextMenuUtils.tsx
@@ -16,7 +16,6 @@ import Wellbore from "models/wellbore";
 import { Fragment } from "react";
 import { RouterLogType } from "routes/routerConstants";
 import {
-  getLogCurveInfoListViewPath,
   getLogObjectViewPath,
   getLogObjectsViewPath,
   getLogTypesViewPath,
@@ -29,7 +28,6 @@ import styled from "styled-components";
 import Icon from "styles/Icons";
 import { openRouteInNewWindow } from "tools/windowHelpers";
 import { ModalContentLayout } from "../StyledComponents/ModalContentLayout";
-import { NavigateFunction, createSearchParams } from "react-router-dom";
 
 export const StyledIcon = styled(Icon)`
   && {
@@ -282,34 +280,3 @@ const displayDeleteModal = (
     payload: confirmation
   });
 };
-
-export const onClickOpenSeveralLogs = (
-  dispatchOperation: DispatchOperation,
-  serverUrl: string,
-  wellUid: string,
-  wellboreUid: string,
-  uid: string,
-  checkedItems: ObjectOnWellbore[],
-  navigate: NavigateFunction,
-  indexCurve: IndexCurve = null
-) => {
-  dispatchOperation({ type: OperationType.HideContextMenu });
-  const path = getLogCurveInfoListViewPath(
-    serverUrl,
-    wellUid,
-    wellboreUid,
-    ObjectType.Log,
-    indexCurve,
-    uid
-  );
-  let searchParams = {};
-  const logUids = getObjects(checkedItems);
-  const logUidsFormatted = JSON.stringify(logUids);
-
-  searchParams = createSearchParams({ logs: logUidsFormatted });
-  navigate({ pathname: path, search: searchParams.toString() });
-};
-
-function getObjects(objectOnWelbore: ObjectOnWellbore[]) {
-  return objectOnWelbore.map((row) => row.uid);
-}
